backend/test_api.py
import requests
import json
import time

BASE_URL = "http://localhost:5000/api"
headers = {"Content-Type": "application/json"}

def print_test(name, response):
    print(f"\n{'='*50}")
    print(f"TEST: {name}")
    print(f"Status: {response.status_code}")
    print(f"Response: {json.dumps(response.json(), indent=2) if response.headers.get('content-type') == 'application/json' else response.text}")
    print(f"Time: {response.elapsed.total_seconds():.3f}s")

def test_health():
    response = requests.get(f"{BASE_URL}/health")
    print_test("Health Check", response)
    return response.status_code == 200

def test_public_endpoints():
    print(f"\n{'='*50}")
    print("TESTING PUBLIC ENDPOINTS")
    print(f"{'='*50}")
    
    response = requests.get(f"{BASE_URL}/public/categories")
    print_test("Public Categories", response)
    
    response = requests.get(f"{BASE_URL}/public/products/featured")
    print_test("Public Featured Products", response)
    
    response = requests.get(f"{BASE_URL}/states")
    print_test("Indian States", response)

def test_user_registration():
    user_data = {
        "email": "test@example.com",
        "password": "password123",
        "first_name": "Test",
        "last_name": "User",
        "phone": "9876543210",
        "referral_code": ""
    }
    
    response = requests.post(f"{BASE_URL}/user/auth/register", 
                           headers=headers, 
                           data=json.dumps(user_data))
    print_test("User Registration", response)
    return response.status_code == 201

def test_user_login():
    login_data = {
        "email": "test@example.com",
        "password": "password123"
    }
    
    response = requests.post(f"{BASE_URL}/user/auth/login",
                           headers=headers,
                           data=json.dumps(login_data))
    print_test("User Login", response)
    
    if response.status_code == 200:
        return response.json().get('token')
    return None

def test_admin_login():
    admin_data = {
        "username": "admin",
        "password": "admin123"
    }
    
    response = requests.post(f"{BASE_URL}/admin/auth/login",
                           headers=headers,
                           data=json.dumps(admin_data))
    print_test("Admin Login", response)
    
    if response.status_code == 200:
        return response.json().get('token')
    return None

def test_user_endpoints(token):
    if not token:
        print("\n⚠️ Skipping user tests - no token")
        return
    
    print(f"\n{'='*50}")
    print("TESTING USER ENDPOINTS")
    print(f"{'='*50}")
    
    auth_headers = {**headers, "Authorization": f"Bearer {token}"}
    
    response = requests.get(f"{BASE_URL}/user/auth/me", headers=auth_headers)
    print_test("Get User Profile", response)
    
    response = requests.get(f"{BASE_URL}/user/products/featured", headers=auth_headers)
    print_test("User Featured Products", response)
    
    response = requests.get(f"{BASE_URL}/user/cart", headers=auth_headers)
    print_test("User Cart", response)
    
    response = requests.get(f"{BASE_URL}/user/referrals", headers=auth_headers)
    print_test("User Referrals", response)

def test_admin_endpoints(token):
    if not token:
        print("\n⚠️ Skipping admin tests - no token")
        return
    
    print(f"\n{'='*50}")
    print("TESTING ADMIN ENDPOINTS")
    print(f"{'='*50}")
    
    auth_headers = {**headers, "Authorization": f"Bearer {token}"}
    
    response = requests.get(f"{BASE_URL}/admin/dashboard", headers=auth_headers)
    print_test("Admin Dashboard", response)
    
    response = requests.get(f"{BASE_URL}/admin/products", headers=auth_headers)
    print_test("Admin Products", response)
    
    response = requests.get(f"{BASE_URL}/admin/users", headers=auth_headers)
    print_test("Admin Users", response)
    
    response = requests.get(f"{BASE_URL}/admin/referrals/stats", headers=auth_headers)
    print_test("Admin Referral Stats", response)

def test_referral_validation():
    print(f"\n{'='*50}")
    print("TESTING REFERRAL SYSTEM")
    print(f"{'='*50}")
    
    test_code_data = {"code": "REFTEST123"}
    response = requests.post(f"{BASE_URL}/user/referrals/validate",
                           headers=headers,
                           data=json.dumps(test_code_data))
    print_test("Validate Referral Code", response)

def test_product_endpoints():
    print(f"\n{'='*50}")
    print("TESTING PRODUCT ENDPOINTS")
    print(f"{'='*50}")
    
    response = requests.get(f"{BASE_URL}/user/products")
    print_test("Get Products", response)
    
    response = requests.get(f"{BASE_URL}/user/products/1")
    print_test("Get Product Detail", response)
    
    response = requests.get(f"{BASE_URL}/user/categories")
    print_test("Get Categories", response)

def run_comprehensive_test():
    print("🚀 Starting Comprehensive API Test Suite")
    print("🔧 Testing New Restructured E-Commerce API")
    print("=" * 70)
    
    start_time = time.time()
    
    test_health()
    test_public_endpoints()
    test_product_endpoints()
    test_referral_validation()
    
    print(f"\n{'='*50}")
    print("TESTING AUTHENTICATION")
    print(f"{'='*50}")
    
    test_user_registration()
    user_token = test_user_login()
    admin_token = test_admin_login()
    
    test_user_endpoints(user_token)
    test_admin_endpoints(admin_token)
    
    total_time = time.time() - start_time
    
    print(f"\n🎉 Test Suite Complete!")
    print("=" * 70)
    print(f"✅ API restructure successful")
    print(f"⚡ Total test time: {total_time:.2f} seconds")
    print(f"🔧 Admin/User separation: Working")
    print(f"🎁 Referral system: Implemented")
    print(f"📊 Caching: Enabled")
    
    print("\n💡 Next steps:")
    print("   - Test with real frontend")
    print("   - Configure email settings")
    print("   - Set up Redis for production")
    print("   - Deploy and test")

if __name__ == "__main__":
    run_comprehensive_test()