backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Caching==2.1.0
mysql-connector-python==8.1.0
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
redis==5.0.1
bcrypt==4.0.1
cryptography==41.0.7
python-dateutil==2.8.2
gunicorn==21.2.0